mercure-pediatric-leglength/requirements.txt
torch>=2.2.1
torchvision>=0.17.1
torchmetrics>=1.7.1
numpy>=1.26.4
pandas>=2.2.1
pydicom>=2.4.4
opencv-python>=4.10.0
matplotlib>=3.8.0
pillow>=10.0.0
requests>=2.28.0
tqdm>=4.67.1
pycocotools>=2.0.8

# Monitoring dependencies (optional)
influxdb-client>=1.37.0
prometheus-client>=0.17.0
psutil>=5.9.0
psycopg2-binary>=2.9.0  # PostgreSQL client for results storage